Add files via upload
The World Health Organization (WHO) has reported that approximately 285 million people are visually impaired worldwide, and out of these 285 million, 39 million are completely blind. It gets extremely tough for them to carry out daily activities, one of which is reading. From reading a newspaper or a magazine to reading an important text message from your bank, it is tough for them to read the text written in it.

 

A similar problem they also face is seeing and enjoying the beauty of pictures and images. Today, in the world of social media, millions of images are uploaded daily. Some of them are about your friends and family, while some of them are about nature and its beauty. Understanding what is present in that image is quite a challenge for certain people who are suffering from visual impairment or who are blind.


In an initiative to help them experience the beauty of the images, Facebook had earlier launched a unique feature earlier that can help blind people operate the app on their mobile phones. The feature could explain the contents of an image that their friends have posted on Facebook. So, say, if someone posted a picture with their dog in the park, the application would speak out the contents and may describe it like, “This image may contain a dog standing with a man around the trees.”


In this module, you will learn how a blind person is able to overcome these problems and carry out activities as others when it comes to seeing. You will learn how to make a model that is similar to the one developed by Facebook, specifically such that a blind person knows the contents of an image in front of them with the help of a CNN-RNN based model. The model will convert the contents of an image and will provide the output in the form of audio.
In this project, you will convert an image to text description first; then, using a simple text-to-speech API, you will extract the text description/caption and convert it to audio. So the central part of this capstone is focused on building the caption/text description whereas the second part, which is transforming the text to speech is relatively easy with the text to speech API. Once the model is built, you will deploy the project on your local system using a Flask-based model to generate audio-based content for any image.

 

Let us see how we can implement this project when deployed using a Flask-based model.
The module consists of four sessions. The first one deals with the concept and foundations of attention models. Once you have understood the theoretical concept of the attention mechanism, you will proceed to the second session that will help you to build custom models using Tensorflow. The third session includes the project, its evaluation and its submission. In the final session, the solution for the project will be provided once the deadline is over.